feat: searchAfter를 쓰는 Book 검색 service와 controller 연결

+++ src/main/java/com/bookspot/book/presentation/response/BookPreviewSearchAfterResponse.java
@@ -6,6 +6,6 @@ public record BookPreviewSearchAfterResponse(
         List<BookPreviewResponse> books,
         Integer lastLoanCount,
         Long lastBookId,
-        int totalElements
+        long totalElements
 ) {
 }
